# services/rest_user_service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/auth/issue": {
            "post": {
                "description": "Выдает Access и Refresh токены для пользователя с указанным GUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Выдача пары токенов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GUID пользователя",
                        "name": "guid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успех",
                        "schema": {
                            "$ref": "#/definitions/typescore.TokenPair"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "description": "Обновляет Access и Refresh токены на основе действующего Refresh токена",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Обновление токенов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer \u003crefresh_token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успех",
                        "schema": {
                            "$ref": "#/definitions/typescore.TokenPair"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Невалидный или просроченный токен",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/users/profile": {
            "get": {
                "description": "Получение профиля пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Получение профиля пользователя",
                "responses": {
                    "200": {
                        "description": "Успех",
                        "schema": {
                            "$ref": "#/definitions/typescore.User"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error_code": {
                    "description": "Код ошибки",
                    "type": "integer"
                },
                "error_description": {
                    "description": "Описание ошибки",
                    "type": "string"
                }
            }
        },
        "typescore.TokenPair": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "typescore.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Дата и время создания записи",
                    "type": "string"
                },
                "email": {
                    "description": "Адрес электронной почты пользователя",
                    "type": "string"
                },
                "first_name": {
                    "description": "Имя пользователя",
                    "type": "string"
                },
                "is_blocked": {
                    "description": "Залочен ли пользователь(заблокирован или нет)",
                    "type": "boolean"
                },
                "last_name": {
                    "description": "Фамилия пользователя",
                    "type": "string"
                },
                "nickname": {
                    "description": "Псевдоним или никнейм пользователя",
                    "type": "string"
                },
                "notification_enabled": {
                    "description": "Включены ли разрешения на push-уведомления",
                    "type": "boolean"
                },
                "role": {
                    "description": "Роль пользователя",
                    "allOf": [
                        {
                            "$ref": "#/definitions/typescore.UserRoleTypes"
                        }
                    ]
                },
                "serial_id": {
                    "description": "Уникальный порядковый идентификатор записи",
                    "type": "integer"
                },
                "system_id": {
                    "description": "Системный идентификатор записи",
                    "type": "string"
                },
                "telegram_id": {
                    "description": "Идентификатор пользователя в Telegram",
                    "type": "integer"
                }
            }
        },
        "typescore.UserRoleTypes": {
            "type": "string",
            "enum": [
                "user",
                "admin",
                "super_admin",
                "support"
            ],
            "x-enum-comments": {
                "AdminRole": "администратор",
                "SuperAdminRole": "супер администратор",
                "SupportRole": "поддержка",
                "UserRole": "пользователь"
            },
            "x-enum-varnames": [
                "UserRole",
                "AdminRole",
                "SuperAdminRole",
                "SupportRole"
            ]
        }
    }
}